requirements.txt
torch
torchvision
diffusers
transformers
safetensors
numpy
Pillow
pytest

